Bug 30952: (follow-up) Restore DataTable sorting icons
Some CSS was removed which controlled the appearance of DataTables
column sort icons. Without it, DataTables tries to use some default
images which do not exist.

This patch restores the icons. With it, sorting icons should look like
they do in master and there should be no related 404 errors in the
console.

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Owen Leonard <oleonard@myacpl.org>
Signed-off-by: Nick Clemens <nick@bywatersolutions.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>