Bug 31635: Fix output of current title in OPAC browse results
This patch makes a correction to the output of OPAC browse results. In
the numbered result navigation links, a title tag on each
navigation link should contain the title of the record. The link for the
current title didn't include the title information in the title tag
because it used the wrong template variables.

To test, apply the patch and make sure OpacBrowseResults is enabled.

- In OPAC, perform a search that will return multiple results.
- View one of the results.
- In the right-hand sidebar, click "Browse results."
- In the list of numbered links at the top of the panel, hover your
  mouse over each link. They should each display a tooltip with  "View
  record [title]"
- Confirm that the current title in the list includes the correct
  tooltip.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>