Bug 31763: Add 'page-section' to patron lists page
This patch updates patron lists templates so that the main content of
the page is wrapped with a "page-section" div. This ensures that content
has sufficient contrast. See Bug 31757.

Unrelated change: I wrapped some form controls in a <fieldset> to make
it consistent with other pages and to work better with the updated
markup.

Note: This patch contains indentation changes, so use diff accordingly.

To test, apply the patch and go to Tools -> Patron lists.

- If necessary, create one or more patron lists so that the table
  appears. Confirm that the content area looks correct.
- Click Actions -> Add patrons to confirm that the screen
  looks correct.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>
Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>