Bug 31764: Add 'page-section' to patron clubs page
This patch updates patron clubs template so that the main content of
the page is wrapped with a "page-section" div. This ensures that content
has sufficient contrast. See Bug 31757.

Note: This patch contains indentation changes, so use diff accordingly.

To test, apply the patch and go to Tools -> Patron clubs.

- If necessary, add at least one club template and at least one club
  based on that template.
- Confirm that the display of club templates and clubs looks correct.
- Locate some patrons and add them to a club.
- On the patron clubs page, locate a club with enrollments and click
  Actions -> Enrollments. Confirm that the page looks correct.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>