Bug 31850: Patron import: welcome email option style as list
This patch corrects the markup around the "Send email to new patrons"
checkbox on the patron import page. The list should be an ordered list,
not an unordered list.

To test, apply the patch and go to Tools -> Import patrons.

In the "Welcome email" section, there should be no bullet point before
the "Send email to new patrons" line.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>