Bug 31753: Dialog boxes inside of modals don't seem wide enough
This patch makes some tweaks to the style of Bootstrap modals in the
staff interface, including making a correction to the style of .dialog
<div>s within modals.

The patch makes a correction to the modal markup generated by the guided
reports template for SQL previews so that the modal footer displays
correctly.

The patch also removes CSS which makes links inside headings the same
color as the heading text. The effect of this was to make links
invisible in the headings of hold confirmation modals.

To test, apply the patch and rebuild the staff interface CSS
(https://wiki.koha-community.org/wiki/Working_with_SCSS_in_the_OPAC_and_staff_client).

- Test the hold confirmation modal in check-in:
  - Place a hold on an available title.
  - Check in an item from that record and confirm that the modal
    looks correct. The "Check in message" box should take up the whole
    width of the modal, and the links in the modal header should be the
    standard green.
- Test the SQL preview modal in reports:
  - Go to Reports -> Use saved.
  - Click the actions menu associated with one of your saved reports and
    choose "Preview SQL." The modal should look correct.
- Other modals which might be tested: MARC previews from the
  bibliographic detail page, the cataloging search page, the Z39.50
  search page, etc.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>